# llm-service/app/src/main/java/ar/edu/utn/frc/tup/piv/llm/adapter/in/web/GoldenSetImportController.java
package ar.edu.utn.frc.tup.piv.llm.adapter.in.web;

import ar.edu.utn.frc.tup.piv.llm.application.model.CallerIdentity;
import ar.edu.utn.frc.tup.piv.llm.application.service.GoldenSetImportService; import ar.edu.utn.frc.tup.piv.llm.domain.goldenset.ImportBatch; import ar.edu.utn.frc.tup.piv.llm.adapter.in.web.security.*; import com.fasterxml.jackson.databind.JsonNode; import java.util.*; import org.springframework.http.*; import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("${app.api.private-path}/courses/{courseId}/golden-set-imports") public class GoldenSetImportController { private final GoldenSetImportService service; private final GoldenSetAuthorization identity; private final CourseAuthorization courses; public GoldenSetImportController(GoldenSetImportService s,GoldenSetAuthorization i,CourseAuthorization c){service=s;identity=i;courses=c;} @PostMapping public ResponseEntity<ImportBatch> create(@PathVariable UUID courseId,@RequestBody ImportRequest r,@RequestHeader("Idempotency-Key") UUID key,@RequestHeader HttpHeaders h){var a=auth(courseId,h);return ResponseEntity.status(201).body(service.create(courseId,r.goldenSetVersionId(),r.format(),r.rows(),key,a.delegatedUserId()));} @PostMapping("/{batchId}/validate") public ImportBatch validate(@PathVariable UUID courseId,@PathVariable UUID batchId,@RequestHeader HttpHeaders h){auth(courseId,h);return service.validate(courseId,batchId);} @PatchMapping("/{batchId}/rows/{rowNumber}") public void replace(@PathVariable UUID courseId,@PathVariable UUID batchId,@PathVariable int rowNumber,@RequestBody JsonNode p,@RequestHeader HttpHeaders h){auth(courseId,h);service.replaceRow(courseId,batchId,rowNumber,p);} @PostMapping("/{batchId}/commit") public ResponseEntity<Void> commit(@PathVariable UUID courseId,@PathVariable UUID batchId,@RequestHeader HttpHeaders h){auth(courseId,h);service.commit(courseId,batchId);return ResponseEntity.status(201).build();} private CallerIdentity auth(UUID c,HttpHeaders h){var a=identity.require(h);courses.requireTeacher(c,a,h);return a;} public record ImportRequest(UUID goldenSetVersionId,String format,List<JsonNode> rows){} }