# llm-service/docs/contracts/llm-service.openapi.yaml
openapi: 3.1.0
info:
  title: LLM Service HTTP API
  version: 2.0.0
  license: {name: UNLICENSED, identifier: UNLICENSED}
  description: Contrato HTTP canónico de la implementación actual. Evalúa uso de IA; no expone corrección académica ni soluciones esperadas.
servers:
  - url: /api/llm
security:
  - serviceJwt: []
paths:
...
  /courses/{courseId}/golden-set-imports/{batchId}/validate:
    post:
      summary: Validate Golden Set Import
      operationId: validateGoldenSetImport
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/BatchId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {$ref: '#/components/responses/ImportBatch'}, '403': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-set-imports/{batchId}/rows/{rowNumber}:
    patch:
      summary: Replace a staged Golden Set import row
      operationId: replaceGoldenSetImportRow
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/BatchId'}, {name: rowNumber, in: path, required: true, schema: {type: integer, minimum: 1}}]
      requestBody: {$ref: '#/components/requestBodies/GoldenSetCaseInput'}
      responses: {'204': {description: Fila reemplazada}, '403': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
  /courses/{courseId}/golden-set-imports/{batchId}/commit:
    post:
      summary: Commit Golden Set Import
      operationId: commitGoldenSetImport
      parameters: [{$ref: '#/components/parameters/CourseId'}, {$ref: '#/components/parameters/BatchId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/ImportBatch'}, '409': {$ref: '#/components/responses/Problem'}}
...
  /model-assignments/{function}:
    get:
      summary: Get the deployment assigned to an internal LLM function
      operationId: getModelAssignment
      parameters: [{name: function, in: path, required: true, schema: {type: string}}]
      responses: {'200': {$ref: '#/components/responses/ApiObject'}, '403': {$ref: '#/components/responses/Problem'}}
    put:
      summary: Assign a deployment to an internal LLM function
      operationId: putModelAssignment
      parameters: [{name: function, in: path, required: true, schema: {type: string}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [modelDeploymentId], properties: {modelDeploymentId: {type: string, format: uuid}}}}}}
      responses: {'200': {$ref: '#/components/responses/ApiObject'}, '403': {$ref: '#/components/responses/Problem'}}
...
  /evaluator-skills:
    get:
      summary: List the evaluator skills catalog
      operationId: listEvaluatorSkillCatalog
      responses:
        '200': { description: Catálogo de skills del evaluador, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EvaluatorSkillResponse' } } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /courses/{courseId}/evaluator-skills:
    get:
      summary: List the evaluator skills of a course with their activation state
      operationId: listCourseEvaluatorSkills
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200': { description: Skills del curso con su estado, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EvaluatorSkillResponse' } } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
    put:
      summary: Configure the active evaluator skills of a course
      operationId: updateCourseEvaluatorSkills
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateCourseSkillsRequest' } } }
      responses:
        '200': { description: Skills del curso luego de la actualización, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/EvaluatorSkillResponse' } } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    serviceJwt: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
...
    CourseId: {name: courseId, in: path, required: true, schema: {type: string, format: uuid}}
...
    BatchId: {name: batchId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
...
    GoldenSetCaseInput: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GoldenSetCaseInput'}}}}
...
  responses:
    Problem: {description: RFC 7807 Problem Details}
...
    ImportBatch: {description: Lote, content: {application/json: {schema: {$ref: '#/components/schemas/ImportBatch'}}}}
...
    ApiObject: {description: Recurso administrativo; sus campos concretos se describen en la operación y en el consumidor de referencia, content: {application/json: {schema: {$ref: '#/components/schemas/ApiObject'}}}}
  schemas:
    Dimensions:
      type: object
      required: [AUTONOMY, CLARITY, PROGRESSION, COMPLIANCE, EFFICIENCY]
      properties:
        AUTONOMY: {type: integer, minimum: 0, maximum: 100}
        CLARITY: {type: integer, minimum: 0, maximum: 100}
        PROGRESSION: {type: integer, minimum: 0, maximum: 100}
        COMPLIANCE: {type: integer, minimum: 0, maximum: 100}
        EFFICIENCY: {type: integer, minimum: 0, maximum: 100}
...
    TranscriptMessage:
      type: object
      required: [role, content, position]
      properties: {role: {type: string, enum: [STUDENT, TUTOR]}, content: {type: string, minLength: 1}, position: {type: integer, minimum: 0}, occurredAt: {type: string, format: date-time}}
    GoldenSetCaseInput:
      type: object
      required: [transcript, referenceScores]
      properties: {transcript: {type: array, minItems: 1, items: {$ref: '#/components/schemas/TranscriptMessage'}}, challengeContext: {type: object, required: [externalChallengeId, statement], properties: {externalChallengeId: {type: string}, statement: {type: string}}}, metadata: {type: object, additionalProperties: true}, author: {type: string, deprecated: true, description: 'Ignorado: el servidor registra al docente autenticado.'}, referenceScores: {$ref: '#/components/schemas/Dimensions'}, scoreJustifications: {type: object, additionalProperties: {type: string}}}
...
    ImportBatch:
      type: object
      required: [id, goldenSetVersionId, format, state, rows]
      properties: {id: {type: string, format: uuid}, goldenSetVersionId: {type: string, format: uuid}, format: {type: string, enum: [JSON, CSV]}, state: {type: string, enum: [DRAFT, VALIDATING, READY, COMMITTED, FAILED]}, rows: {type: integer, minimum: 0, description: 'Cantidad de filas cargadas. El detalle se consulta/edita por la API de staging.'}}
...
    ApiObject: {type: object, additionalProperties: true}
...
    EvaluatorSkillResponse:
      type: object
      required: [skillKey, name, description, toolType, isActive]
      properties:
        skillKey: {type: string, description: Clave canónica de la skill en el catálogo.}
        name: {type: string}
        description: {type: string}
        toolType: {type: string, description: Tipo de herramienta con que se evalúa la skill.}
        isActive: {type: boolean, description: Habilitada por defecto (catálogo) o activada para el curso.}
    UpdateCourseSkillsRequest:
      type: object
      required: [activeSkillKeys]
      properties:
        activeSkillKeys: {type: array, items: {type: string}, description: Claves del catálogo que quedan activas para el curso. Reemplaza el set anterior.}